fixed signed/unsigned comparison warning
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@46394 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/wxcrt.cpp
@@ -576,7 +576,7 @@ int wxCRT_VsnprintfA(char *buf, size_t len, const char *format, va_list argptr)
     int rt = wxCRT_VsnprintfW(wbuf.data(), len,
                               (const wchar_t*)wxConvLibc.cMB2WC(format),
                               argptr);
-    if ( rt < 0 || rt >= len )
+    if ( rt < 0 || rt >= (int)len )
         return rt;
 
     if ( wxConvLibc.FromWChar(buf, len, wbuf) == wxCONV_FAILED )
